Themes: Use SRC_ATOP tintMode for external tile icons
For more elaborate themes that use a drawable with color and no
need for tinting, the tint mode needs to be changed to SRC_ATOP
instead of the default SRC_IN.  This means that the alpha value
from the tint color is discarded and the alpha of the drawable is
used instead.  Stock still looks as expected so this should be
an acceptable compromise to accommodate themes that supply a
transparent tint with the intention of no tint being applied.

Change-Id: I6852792b6b71ef73c5c0ba6073c88bff8c19456b