[2/2] Settings: On-The-Go mode
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

have fun!

Change-Id: I8a43f56b96bb1e2fe77bce8b438b18ac7505b2be
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>