SetupWizard: fix Vision Settings crash
Google's SetupWizard assumes that there is always going to be a Talbback/ScreenReader package in the ROM.
This isn't the case with many GApps packages, so Vision Settings FC.

Also, by looking at the code, it feels like they had actually thought about this case, and they decided to use a null check.
It was left unfinished though, because the same null value is getting used outside the null check (hence the FC).

This commit fixes it by removing the preference if a package doesn't exist, and properly applying the null check.
Or you can bundle Google Talkback in every GApps package.

Change-Id: I979be1c91f4ec36aac987d412b7df91e1810cc2f
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>